Cleaned ExcelDna.ComInterop files.
