+++ Source/ExcelDna.Integration/ComInterop/ComInterop.cs
@@ -7,14 +7,13 @@ using System.Runtime.InteropServices.ComTypes;
 using ExcelDna.ComInterop;
 
 using HRESULT = System.Int32;
-using IID = System.Guid;
-using CLSID = System.Guid;
-using DWORD = System.Int32;
+using IID     = System.Guid;
+using CLSID   = System.Guid;
+using DWORD   = System.Int32;
 using System.Runtime.CompilerServices;
 using System.Reflection;
 using System.Collections;
 using System.Threading;
-using ExcelDna.Integration;
 
 namespace ExcelDna.ComInterop
 {
@@ -174,11 +173,11 @@ namespace ExcelDna.Integration.CustomUI
     public interface IRibbonControl
     {
         [DispId(1)]
-        string Id { [return: MarshalAs(UnmanagedType.BStr)][DispId(1)] get; }
+        string Id { [return: MarshalAs(UnmanagedType.BStr)] [DispId(1)] get; }
         [DispId(2)]
-        object Context { [return: MarshalAs(UnmanagedType.IDispatch)][DispId(2)] get; }
+        object Context { [return: MarshalAs(UnmanagedType.IDispatch)] [DispId(2)] get; }
         [DispId(3)]
-        string Tag { [return: MarshalAs(UnmanagedType.BStr)][DispId(3)] get; }
+        string Tag { [return: MarshalAs(UnmanagedType.BStr)] [DispId(3)] get; }
     }
 
     [ComImport]
@@ -192,14 +191,14 @@ namespace ExcelDna.Integration.CustomUI
         [DispId(2)]
         void InvalidateControl([In, MarshalAs(UnmanagedType.BStr)] string ControlID);
         [DispId(3)]
-        void InvalidateControlMso([MarshalAs(UnmanagedType.BStr)][In] string ControlID);
+        void InvalidateControlMso([MarshalAs(UnmanagedType.BStr)] [In] string ControlID);
         // The ActiveTab methods were added in Office 2010
         [DispId(4)]
-        void ActivateTab([MarshalAs(UnmanagedType.BStr)][In] string ControlID);
+        void ActivateTab([MarshalAs(UnmanagedType.BStr)] [In] string ControlID);
         [DispId(5)]
-        void ActivateTabMso([MarshalAs(UnmanagedType.BStr)][In] string ControlID);
+        void ActivateTabMso([MarshalAs(UnmanagedType.BStr)] [In] string ControlID);
         [DispId(6)]
-        void ActivateTabQ([MarshalAs(UnmanagedType.BStr)][In] string ControlID, [MarshalAs(UnmanagedType.BStr)][In] string Namespace);
+        void ActivateTabQ([MarshalAs(UnmanagedType.BStr)] [In] string ControlID, [MarshalAs(UnmanagedType.BStr)] [In] string Namespace);
     }
 
     //// Actually from System.Windows.Forms.UnsafeNativeMethods
@@ -249,23 +248,23 @@ namespace ExcelDna.Integration.CustomUI
     public interface ICustomTaskPane
     {
         [DispId(0)]
-        string Title { [return: MarshalAs(UnmanagedType.BStr)][MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(0)] get; }
+        string Title { [return: MarshalAs(UnmanagedType.BStr)] [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(0)] get; }
         [DispId(1)]
-        object Application { [return: MarshalAs(UnmanagedType.IDispatch)][MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(1)] get; }
+        object Application { [return: MarshalAs(UnmanagedType.IDispatch)] [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(1)] get; }
         [DispId(2)]
-        object Window { [return: MarshalAs(UnmanagedType.IDispatch)][MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(2)] get; }
+        object Window { [return: MarshalAs(UnmanagedType.IDispatch)] [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(2)] get; }
         [DispId(3)]
-        bool Visible { [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(3)] get; [param: In][MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(3)] set; }
+        bool Visible { [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(3)] get; [param: In] [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(3)] set; }
         [DispId(4)]
-        object ContentControl { [return: MarshalAs(UnmanagedType.IDispatch)][MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(4)] get; }
+        object ContentControl { [return: MarshalAs(UnmanagedType.IDispatch)] [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(4)] get; }
         [DispId(5)]
-        int Height { [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(5)] get; [param: In][MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(5)] set; }
+        int Height { [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(5)] get; [param: In] [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(5)] set; }
         [DispId(6)]
-        int Width { [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(6)] get; [param: In][MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(6)] set; }
+        int Width { [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(6)] get; [param: In] [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(6)] set; }
         [DispId(7)]
-        MsoCTPDockPosition DockPosition { [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(7)] get; [param: In][MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(7)] set; }
+        MsoCTPDockPosition DockPosition { [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(7)] get; [param: In] [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(7)] set; }
         [DispId(8)]
-        MsoCTPDockPositionRestrict DockPositionRestrict { [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(8)] get; [param: In][MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(8)] set; }
+        MsoCTPDockPositionRestrict DockPositionRestrict { [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(8)] get; [param: In] [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(8)] set; }
         [MethodImpl(MethodImplOptions.InternalCall, MethodCodeType = MethodCodeType.Runtime), DispId(9)]
         void Delete();
     }
@@ -597,7 +596,7 @@ namespace ExcelDna.Integration.Rtd
         [DispId(12)]
         [return: MarshalAs(UnmanagedType.SafeArray, SafeArraySubType = VarEnum.VT_VARIANT)]
         Array RefreshData(ref int topicCount);
-
+        
         //
         // Summary:
         //     Notifies a real-time data (RTD) server application that a topic is no longer
@@ -848,72 +847,3 @@ namespace ExcelDna.ComInterop
 
 }
 #endregion
-
-namespace ExcelDna.ComInterop
-{
-    public interface IType
-    {
-        object GetObject(IntPtr pUnk);
-        void ReleaseObject(object comObject);
-
-        bool HasProperty(string name, object comObject);
-        object GetProperty(string name, object comObject);
-        void SetProperty(string name, object value, object comObject);
-        object GetIndex(int i, object comObject);
-        object Invoke(string name, object[] args, object comObject);
-        bool Is(ref Guid guid, object comObject);
-    }
-
-    internal static class Util
-    {
-        public static IType TypeAdapter
-        { get; } = NativeAOT.IsActive ? NativeAOT.TypeAdapter : new TypeAdapter();
-    }
-
-    internal class TypeAdapter : IType
-    {
-        public object GetProperty(string name, object comObject)
-        {
-            return comObject.GetType().InvokeMember(name, BindingFlags.GetProperty, null, comObject, null);
-        }
-
-        public object GetIndex(int i, object comObject)
-        {
-            return comObject.GetType().InvokeMember("", BindingFlags.GetProperty, null, comObject, new object[] { i });
-        }
-
-        public bool Is(ref CLSID guid, object comObject)
-        {
-            IntPtr pUnk = Marshal.GetIUnknownForObject(comObject);
-
-            IntPtr pButton;
-            Marshal.QueryInterface(pUnk, ref guid, out pButton);
-            return (pButton != IntPtr.Zero);
-        }
-
-        public object Invoke(string name, object[] args, object comObject)
-        {
-            return comObject.GetType().InvokeMember(name, BindingFlags.InvokeMethod, null, comObject, args);
-        }
-
-        public void SetProperty(string name, object value, object comObject)
-        {
-            comObject.GetType().InvokeMember(name, BindingFlags.SetProperty, null, comObject, new object[] { value });
-        }
-
-        public object GetObject(IntPtr pUnk)
-        {
-            return Marshal.GetObjectForIUnknown(pUnk);
-        }
-
-        public void ReleaseObject(object comObject)
-        {
-            Marshal.ReleaseComObject(comObject);
-        }
-
-        public bool HasProperty(string name, object comObject)
-        {
-            return ComInterop.DispatchHelper.HasProperty(comObject, name);
-        }
-    }
-}
