+++ Source/ExcelDna.Integration/ComInterop/ComRegistration.cs
