+++ Source/ExcelDna.Integration/ComInterop/IType.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace ExcelDna.Integration.ComInterop
+{
+    public interface IType
+    {
+        object GetObject(IntPtr pUnk);
+        void ReleaseObject(object comObject);
+
+        bool HasProperty(string name, object comObject);
+        object GetProperty(string name, object comObject);
+        void SetProperty(string name, object value, object comObject);
+        object GetIndex(int i, object comObject);
+        object Invoke(string name, object[] args, object comObject);
+        bool Is(ref Guid guid, object comObject);
+    }
+}
